+++ apps/opentelemetry/test/otel_sweeper_SUITE.erl
@@ -10,11 +10,11 @@
 -include_lib("common_test/include/ct.hrl").
 
 -include("otel_test_utils.hrl").
--include("otel_span.hrl").
+-include_lib("opentelemetry_sdk/include/otel_span.hrl").
 -include_lib("opentelemetry_api/include/opentelemetry.hrl").
 -include_lib("opentelemetry_api/include/otel_tracer.hrl").
 
--include("../src/otel_span_ets.hrl").
+-include_lib("opentelemetry_sdk/src/otel_span_ets.hrl").
 
 all() ->
     [storage_size,