+++ apps/opentelemetry_experimental/src/otel_otlp_logs.erl
@@ -26,7 +26,7 @@
 
 -include_lib("kernel/include/logger.hrl").
 -include_lib("opentelemetry_api/include/opentelemetry.hrl").
--include_lib("opentelemetry/include/otel_span.hrl").
+-include_lib("opentelemetry_sdk/include/otel_span.hrl").
 
 -define(IS_STRING(String),
         (is_list(String) orelse is_binary(String))).