+++ apps/opentelemetry_sdk/src/otel_exporter.erl
@@ -97,7 +97,7 @@ init({ExporterModule, Config}) when is_atom(ExporterModule) ->
                     end;
                 {error, undef} when ExporterModule =:= opentelemetry_exporter ->
                     ?LOG_WARNING("OTLP exporter module `opentelemetry_exporter` not found. "
-                                 "Verify you have included the `opentelemetry_exporter` dependency."),
+                                 "Verify you have included the `opentelemetry_sdk` dependency."),
                     undefined;
                 {error, undef} ->
                     ?LOG_WARNING("Exporter module ~tp not found. Verify you have included "